fix(e2e): move techdocs files under overlay/src from overlay
